move confluence-specific config exception into a new file
Moving the exception `ConfluenceConfigurationError` into its own
exception file under `config`. This is to prepare for updates to
configuration exceptions, where various exception cases will have their
own class instead of custom messages defined in scenarios like the
configuration check call.
